Android: Seperate GameAdapter class, changed Async class to normal class
GameAdapter Seperated to make it importable into an Async
Async class changed, so that those functions can be called from other Async classes
